рефакторинг бебиблиотеки пула объектов: 1. Изменил иерархию наследования объекто 2. Сделал привязку пула объектов к демодулятору

+++ src/include/objects/base/GraphicsObject.cpp
@@ -13,7 +13,7 @@ GraphicsObject::~GraphicsObject()
 
 }
 
-void GraphicsObject::setRotateAngle(double rta)
+void GraphicsObject::setRotateAngle(float rta)
 {
     _rtAngle = rta;
 }
